Add RequesterID support in the Scoping section
The SAML specification supports the RequesterID element to identify the set of requesting entities
on whose behalf the requester is acting. Used to communicate the chain of requesters when proxying
occurs. The org.opensaml:opensaml:2.6.1 library supports this, but there was no option in WebSSOProfileOptions
to set the RequesterID.

See also: https://jira.spring.io/browse/SES-171