Persist the logout request before sending it to the user
We're currently suffering from a race condition where SAML crashes with
a white page on being redirected back from the IdP when logging out.

This is because our session persistence mechanism can take longer
to persist session information than our IdP can take to log out.

Spring-SAML checks that the response to a logout request was for a
request that it created, and throws if this is not the case.

With the current implementation, this means that we are seeing the
following flow.

1. We redirect the user to the IdP
2. We start persisting the request.
3. The user returns from the IdP and we don't find the request.
4. We finish persisting the request.

With this PR, we change the flow to

1. We persist the request.
2. We redirect the user to the IdP.
3. The user returns from the IdP and we find the request.